Starts working on test suite
