+++ egon/mock.py
@@ -4,11 +4,13 @@ for a pre-defined number of seconds.
 """
 
 from egon import nodes
-from egon._utils import MPool
+from egon.utils import MPool
 from egon.connectors import Input, Output
 
 
 class Mock:
+    """Base class for mock testing nodes"""
+
     def run_mock(self) -> None:
         self._pool: MPool
         self._pool.start()
