+++ egon/nodes.py
@@ -6,15 +6,17 @@ that produce, analyze, and consume data respectively.
 from __future__ import annotations
 
 import abc
+import multiprocessing as mp
 from abc import ABC
 from itertools import chain
-from typing import Collection, List, Tuple, Union
+from time import sleep
+from typing import Collection
+from typing import List, Tuple, Union
 
 from . import connectors, exceptions
-from ._utils import MPool
 
 
-def _get_nodes_from_connectors(connector_list: Collection[connectors.AbstractConnector]) -> Tuple:
+def _get_nodes_from_connectors(connector_list: Collection[connectors.BaseConnector]) -> Tuple:
     """Return the parent nodes from a list of ``Connector`` objects
     
     Args:
@@ -25,6 +27,72 @@ def _get_nodes_from_connectors(connector_list: Collection[connectors.AbstractCon
     """
 
     return tuple(p.parent_node for c in connector_list for p in c.partners)
+
+
+class MPool:
+    """A pool of processes assigned to a single target function"""
+
+    def __init__(self, num_processes: int, target: callable) -> None:
+        """Create a collection of processes assigned to execute a given callable
+
+        Args:
+            num_processes: The number of processes to allocate
+            target: The function to be executed by the allocated processes
+        """
+
+        if num_processes <= 0:
+            raise ValueError(f'Cannot instantiate less than 1 forked processes (got {num_processes}).')
+
+        # Note that we use the memory address of the processes and not the
+        # ``pid`` attribute. ``pid`` is only set after the process is started.
+        self._processes = [mp.Process(target=self._call_target) for _ in range(num_processes)]
+        self._states = mp.Manager().dict({id(p): False for p in self._processes})
+        self._target = target
+
+    def _call_target(self) -> None:  # pragma: nocover, Called from forked process
+        """Wrapper for calling the pool's target function"""
+
+        self._target()
+
+        # Mark the current process as being finished
+        sleep(.5)  # Allow any last minute calls to finish before changing the state
+        self._states[id(mp.current_process())] = True
+
+    @property
+    def num_processes(self) -> int:
+        """The number of processes assigned to the pool"""
+
+        return len(self._processes)
+
+    @property
+    def target(self) -> callable:
+        """The callable to be executed by the pool"""
+
+        return self._target
+
+    def is_finished(self) -> bool:
+        """Return whether all processes have finished executing"""
+
+        # Check that all forked processes are finished
+        return all(self._states.values())
+
+    def start(self) -> None:
+        """Start all processes asynchronously"""
+
+        for p in self._processes:
+            p.start()
+
+    def join(self) -> None:
+        """Wait for any running pool processes to finish running before continuing execution"""
+
+        for p in self._processes:
+            p.join()
+
+    def kill(self) -> None:
+        """Kill all running processes without trying to exit gracefully"""
+
+        for p in self._processes:
+            p.terminate()
 
 
 class AbstractNode(abc.ABC):
@@ -39,7 +107,7 @@ class AbstractNode(abc.ABC):
 
         self._inputs = []
         self._outputs = []
-        for connector in self._get_attrs(connectors.AbstractConnector):
+        for connector in self._get_attrs(connectors.BaseConnector):
             connector._node = self  # Make the connector aware of its parent node
             if isinstance(connector, connectors.Input):
                 self._inputs.append(connector)
@@ -146,13 +214,13 @@ class AbstractNode(abc.ABC):
         for input_connector in self._get_attrs(connectors.Input):
             # IMPORTANT: The order of the following code blocks is crucial
             # We check for any running upstream nodes first
-            for output_connector in input_connector.get_partners():
-                if not output_connector.parent_node.pool_finished:
+            for output_connector in input_connector.partners:
+                if not output_connector.parent_node.is_finished():
                     return True
 
             # Check for any unprocessed data once we know there are no
             # nodes still populating any input queues
-            if not input_connector.empty():
+            if not input_connector.is_empty():
                 return True
 
         return False
