+++ egon/pipeline.py
@@ -118,8 +118,8 @@ class Pipeline:
 
     def visualize(
             self,
-            host: str = os.getenv("HOST", "127.0.0.1"),
-            port: int = os.getenv("PORT", "8050"),
+            host: str = os.getenv("EGON_HOST", "127.0.0.1"),
+            port: int = os.getenv("EGON_PORT", "8050"),
             quiet: bool = False
     ) -> None:
         """Launch a server instance for monitoring the pipeline in real time
