+++ tests/test_nodes/test_generic.py
@@ -43,12 +43,12 @@ class TreeNavigation(TestCase):
     def test_upstream_nodes(self) -> None:
         """Test the inline node resolves the correct parent node"""
 
-        self.assertEqual(self.root, self.internal_node.upstream_nodes()[0])
+        self.assertEqual(self.root, self.internal_node.upstream_nodes[0])
 
     def test_downstream_nodes(self) -> None:
         """Test the inline node resolves the correct child node"""
 
-        self.assertEqual(self.leaf, self.internal_node.downstream_nodes()[0])
+        self.assertEqual(self.leaf, self.internal_node.downstream_nodes[0])
 
 
 class ExpectingData(TestCase):
@@ -70,24 +70,24 @@ class ExpectingData(TestCase):
         """Test the return is False for a EMPTY queue and a FINISHED PARENT node"""
 
         self.root._process_finished = True
-        self.assertFalse(self.node.is_expecting_data)
+        self.assertFalse(self.node.is_expecting_data())
 
     def test_true_if_input_queue_has_data(self) -> None:
         """Test the return is True for a NOT EMPTY queue and a FINISHED PARENT node"""
 
         self.root._process_finished = True
         self.node.input._queue.put(5)
-        self.assertTrue(self.node.is_expecting_data)
+        self.assertTrue(self.node.is_expecting_data())
 
     def test_true_if_parent_is_running(self) -> None:
         """Test the return is True for a EMPTY queue and a NOT FINISHED PARENT node"""
 
         self.root._process_finished = False
-        self.assertTrue(self.node.is_expecting_data)
+        self.assertTrue(self.node.is_expecting_data())
 
     def test_true_if_input_queue_has_data_and_parent_is_running(self) -> None:
         """Test the return is True for a NOT EMPTY queue and a NOT FINISHED PARENT node"""
 
         self.root._process_finished = False
         self.node.input._queue.put(5)
-        self.assertTrue(self.node.is_expecting_data)
+        self.assertTrue(self.node.is_expecting_data())
