+++ tests/test_nodes/test_mpool.py
@@ -1,10 +1,12 @@
 from time import sleep
 from unittest import TestCase
 
-from egon._utils import MPool
+from egon.utils import MPool
 
 
 def target_func() -> None:
+    """A dummy target function"""
+
     sleep(2)
 
 
@@ -21,29 +23,11 @@ class ProcessAllocation(TestCase):
         self.assertEqual(self.num_processes, self.pool.num_processes)
         self.assertEqual(self.num_processes, len(self.pool._processes))
 
-    def test_reallocation(self) -> None:
-        """Test a new set of processes are allocated when the number of processes is changed"""
-
-        new_process_count = 2
-        self.pool.num_processes = new_process_count
-        self.assertEqual(new_process_count, self.pool.num_processes)
-        self.assertEqual(new_process_count, len(self.pool._processes))
-
-    def test_error_if_processes_are_alive(self) -> None:
-        """Test a RuntimeError is raised when trying to reallocate processes on a running node"""
-
-        self.pool.start()
-        with self.assertRaises(RuntimeError):
-            self.pool.num_processes = 1
-
     def test_error_on_negative_processes(self) -> None:
         """Assert a value error is raised when the ``num_processes`` attribute is set to a negative"""
 
         with self.assertRaises(ValueError):
             MPool(-1, target_func)
-
-        with self.assertRaises(ValueError):
-            self.pool.num_processes = -1
 
     def test_error_on_zero_processes(self) -> None:
         """Assert a value error is raised when the ``num_processes`` attribute is set to zero"""
@@ -51,22 +35,19 @@ class ProcessAllocation(TestCase):
         with self.assertRaises(ValueError):
             MPool(0, target_func)
 
-        with self.assertRaises(ValueError):
-            self.pool.num_processes = 0
-
 
 class Execution(TestCase):
 
     def test_pool_is_finished_after_execution(self) -> None:
-        """Test the ``pool_finished`` property is updated after the pool executes"""
+        """Test the ``is_finished`` property is updated after the pool executes"""
 
         pool = MPool(2, target_func)
-        self.assertFalse(pool.pool_finished, 'Default finished state is not False.')
+        self.assertFalse(pool.is_finished(), 'Default finished state is not False.')
 
         pool.start()
         pool.join()
 
-        self.assertTrue(pool.pool_finished)
+        self.assertTrue(pool.is_finished())
 
     def test_processes_killed_on_command(self) -> None:
         """Test processes are killed on demand"""
@@ -76,5 +57,5 @@ class Execution(TestCase):
         pool.kill()
         pool.join()
 
-        self.assertTrue(pool._processes[0].exitcode < 0, f'Process not ended by termination signal ({pool._processes[0].exitcode })')
-        self.assertFalse(pool.pool_finished)
+        self.assertTrue(pool._processes[0].exitcode < 0, f'Process not ended by termination signal ({pool._processes[0].exitcode})')
+        self.assertFalse(pool.is_finished())
